requirements.txt
requests == 2.18.4
python-telegram-bot == 9.0.0
